+++ .forge.json
@@ -1,0 +1,26 @@
+{
+  "make_targets": {
+    "win32": [
+      "squirrel"
+    ],
+    "darwin": [
+      "dmg"
+    ],
+    "linux": [
+      "deb"
+    ]
+  },
+  "electronInstallerDMG": {
+    "icon": "src/images/icon.icns"
+  },
+  "electronWinstallerConfig": {
+    "name": "translationCore",
+    "setupIcon": "src/images/icon.ico"
+  },
+  "github_repository": {
+    "owner": "klappy",
+    "name": "translationCore",
+    "draft": "true",
+    "prerelease": "true"
+  }
+}
