--- scripts/inno/windows.iss
@@ -1,59 +1,0 @@
-; Script generated by the Inno Script Studio Wizard.
-; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
-
-#define AppName "translationCore"
-#define AppPublisher "unfoldingWord"
-#define Website "https://unfoldingword.org"
-#define AppExe "translationCore.exe"
-#define RootPath "Z:\home\travis\build\unfoldingWord-dev\translationCore\"
-#define BuildPath RootPath + "dist\translationCore-win32-ia32\"
-#define BuildContents BuildPath + "*"
-#define ExecPath BuildPath + "translationCore.exe"
-#define IconPath RootPath + "scripts\builds\assets\icon.ico"
-#define LicenseFile RootPath + "LICENSE"
-#define OutputDir BuildPath + "\installer\"
-#ifndef Version
-  #define Version "0.0.0"
-#endif
-
-[Setup]
-; NOTE: The value of AppId uniquely identifies this application.
-; Do not use the same AppId value in installers for other applications.
-; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
-AppId={{7D9FD2D0-BB0E-40E8-AEED-6FD604BE6E37}
-AppName={#AppName}
-AppVersion={#Version}
-;AppVerName={#AppName} {#Version}
-AppPublisher={#AppPublisher}
-AppPublisherURL={#Website}
-AppSupportURL={#Website}
-AppUpdatesURL={#Website}
-DefaultDirName={pf}\translationCore
-DisableDirPage=yes
-DefaultGroupName={#AppName}
-AllowNoIcons=yes
-OutputBaseFilename=translationCoreSetup
-OutputDir={#OutputDir}
-SetupIconFile={#IconPath}
-Compression=lzma
-SolidCompression=yes
-LicenseFile={#LicenseFile}
-
-[Languages]
-Name: "english"; MessagesFile: "compiler:Default.isl"
-
-[Tasks]
-Name: "desktopicon"; Description: "{cm:CreateDesktopIcon}"; GroupDescription: "{cm:AdditionalIcons}"; Flags: unchecked
-
-[Files]
-Source: {#ExecPath}; DestDir: "{app}"; Flags: ignoreversion
-Source: {#BuildContents}; DestDir: "{app}"; Flags: ignoreversion recursesubdirs createallsubdirs
-; NOTE: Don't use "Flags: ignoreversion" on any shared system files
-
-[Icons]
-Name: "{group}\{#AppName}"; Filename: "{app}\{#AppExe}"
-Name: "{commondesktop}\{#AppName}"; Filename: "{app}\{#AppExe}"; Tasks: desktopicon
-Name: "{group}\{cm:UninstallProgram,{#AppName}}"; Filename: "{uninstallexe}"
-
-[Run]
-Filename: "{app}\{#AppExe}"; Description: "{cm:LaunchProgram,{#StringChange(AppName, '&', '&&')}}"; Flags: nowait postinstall skipifsilent
