+++ src/index.html
@@ -6,18 +6,18 @@
     <title></title>
     <base target="_blank">
     <!-- Latest compiled and minified CSS -->
-    <link rel="stylesheet" href="./node_modules/bootstrap/dist/css/bootstrap.min.css" >
+    <link rel="stylesheet" href="css/bootstrap.min.css" >
 
     <!-- Optional theme -->
-    <link rel="stylesheet" href="./node_modules/bootstrap/dist/css/bootstrap-theme.min.css">
+    <link rel="stylesheet" href="css/bootstrap-theme.min.css">
 
     <!-- Custom CSS -->
-    <link rel="stylesheet" type="text/css" href="src/css/styles.css">
+    <link rel="stylesheet" type="text/css" href="css/styles.css">
   </head>
   <body>
     <div id="content"></div>
     <script type="text/javascript">
-      require("./src/js/pages/index");
+      require("./js/pages/index");
     </script>
   </body>
 </html>
