+++ src/js/components/loader/Loader.js
@@ -11,7 +11,7 @@ class Loader extends React.Component {
       <MuiThemeProvider>
         <Dialog modal={true} open={show}>
           <div style={{height: "500px", display: 'flex', flexDirection: 'column', justifyContent: 'center', alignItems: 'center', padding: "20px"}}>
-            <img className="App-logo" src={window.__base + "src/images/TC_Icon.png"} alt="logo" style={{height: "350px", margin: "15px"}}/>
+            <img className="App-logo" src={window.__base + "images/TC_Icon.png"} alt="logo" style={{height: "350px", margin: "15px"}}/>
             <span style={{margin: "20px"}}>Loading ...</span>
           </div>
         </Dialog>
