+++ src/js/pages/app.js
@@ -17,6 +17,8 @@ import AlertDialogContainer from '../containers/AlertDialogContainer';
 // actions
 import * as ResourcesActions from '../actions/ResourcesActions';
 
+import packageJson from '../../../package.json';
+
 class Main extends Component {
 
   componentWillMount() {
@@ -25,11 +27,11 @@ class Main extends Component {
   }
 
   componentDidMount() {
-    var packageJson = require(window.__base + '/package.json');
     if (localStorage.getItem('version') !== packageJson.version) {
       localStorage.setItem('version', packageJson.version);
     }
-    this.props.actions.getResourcesFromStaticPackage();
+    // TODO: Figure out why this breaks thunkMiddleware
+    // this.props.actions.getResourcesFromStaticPackage();
   }
 
   render() {
